Manage Threads In Context
Joining threads in the destructor for a resource causes intermittent segfaults
because the GC doesn't always run at the right time/location (AFAIK).
Rather than dance around this problem, I've just moved all of the
joining into the Context's descructor which ensures that the threads
join _before_ the module unloads. This seems to fix the constant
segfaulting problem.